// src/nebular-react-styles/src/theme/types/nebular-theme/NebularThemeOption.ts
export interface NebularThemeOption {
  option_list_max_height: any;
  option_list_shadow: any;
  option_list_background_color: any;
  option_list_border_style: any;
  option_list_border_width: any;
  option_list_border_color: any;
  option_list_border_radius: any;
  option_list_adjacent_border_color: any;
  option_list_adjacent_border_style: any;
  option_list_adjacent_border_width: any;

  option_group_text_color: any;
  option_group_tiny_start_padding: any;
  option_group_small_start_padding: any;
  option_group_medium_start_padding: any;
  option_group_large_start_padding: any;
  option_group_giant_start_padding: any;

  option_background_color: any;
  option_text_color: any;
  option_text_font_family: any;
  option_hover_background_color: any;
  option_hover_text_color: any;
  option_active_background_color: any;
  option_active_text_color: any;
  option_focus_background_color: any;
  option_focus_text_color: any;
  option_selected_background_color: any;
  option_selected_text_color: any;
  option_selected_hover_background_color: any;
  option_selected_hover_text_color: any;
  option_selected_active_background_color: any;
  option_selected_active_text_color: any;
  option_selected_focus_background_color: any;
  option_selected_focus_text_color: any;
  option_disabled_background_color: any;
  option_disabled_text_color: any;

  option_list_scrollbar_color: any;
  option_list_scrollbar_background_color: any;
  option_list_scrollbar_width: any;
}